% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaCS_funcs.R
\name{all_compounds_plot}
\alias{all_compounds_plot}
\title{---creates bar plots for all compounds}
\usage{
all_compounds_plot(df, ctrls_avg, comp_types, conc)
}
\description{
---creates bar plots for all compounds
}
