% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaCS_funcs.R
\name{controls_bar}
\alias{controls_bar}
\title{---creates box and bar plots for all controls}
\usage{
controls_bar(df, tbl, conc, conc_st_err, color)
}
\description{
---creates box and bar plots for all controls
}
